% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_net_position.R
\name{avg_net_position}
\alias{avg_net_position}
\title{Average Net Position}
\usage{
avg_net_position(participant, payments, debit, t_start = NULL,
  t_end = NULL)
}
\arguments{
\item{participant}{character value, Identifying code of the payments system
participant}

\item{payments}{dataframe value, payments data}

\item{debit}{logical value, if this value is TRUE then the net debit position
is calculated. If FALSE then the net credit position is calculated.}

\item{t_start}{character value, First time in the day when payments are sent.
Format is HH:MM:SS (ex. 03:09:23)
Defaults to NULL, taking the time of the first payment in the
data as it's value.}

\item{t_end}{character value, Last time in the day when payments are sent. Format
is HH:MM:SS (ex. 03:09:23)
Defaults to NULL, taking the time of the last payment in the
data as it's value.}
}
\value{
dataframe structure, The average net debit/credit position of the given participant
}
\description{
Average Net Position
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
