% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_payment_timing.R
\name{avg_payment_timing}
\alias{avg_payment_timing}
\title{Average Time of Payments Sent}
\usage{
avg_payment_timing(payments)
}
\arguments{
\item{payments}{dataframe value, payments data}
}
\value{
dataframe structure, daily time values when the average payment
        was sent through the system
}
\description{
Average Time of Payments Sent
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
