% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deciles.R
\name{decile}
\alias{decile}
\title{Decile}
\usage{
decile(payments)
}
\arguments{
\item{payments}{dataframe value, payments data}
}
\value{
dataframe structure, daily time values of the deciles of payments
}
\description{
Decile
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to

For each day in payments finds the time when the (0.1, 0.2, 0.3,..., 0.9)
percentiles of payments were made
}
