% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_coefficient.R
\name{gini_coefficient}
\alias{gini_coefficient}
\title{Gini Coefficient}
\usage{
gini_coefficient(payments)
}
\arguments{
\item{payments}{dataframe value, payments data}
}
\value{
dataframe structure, daily gini coefficient of the system
}
\description{
Gini Coefficient
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
