% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_liq_prov.R
\name{max_liq_prov}
\alias{max_liq_prov}
\title{Max Intraday Liquidity Provision}
\usage{
max_liq_prov(participant, payments, debit)
}
\arguments{
\item{participant}{character value, Identifying code of the payments system
participant}

\item{payments}{dataframe value, payments data}

\item{debit}{boolean value, logical value, if this value is TRUE then the net debit position
is calculated. If FALSE then the net credit position is calculated.}
}
\value{
dataframe structure, The max net debit/credit position of the given participant
        in billions
}
\description{
Max Intraday Liquidity Provision
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
