% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_wide_liq_eff.R
\name{sys_wide_liq_eff}
\alias{sys_wide_liq_eff}
\title{System Wide Liquidity Efficiency}
\usage{
sys_wide_liq_eff(payments)
}
\arguments{
\item{payments}{character value, specific payments data frame}
}
\value{
dataframe structure, System wide liquidity efficiency for each day in
        inputed payments
}
\description{
System Wide Liquidity Efficiency
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
