% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_coefficient.R
\name{gini_coefficient}
\alias{gini_coefficient}
\title{Gini Coefficient}
\usage{
gini_coefficient(payments, liquidity_provided)
}
\arguments{
\item{payments}{dataframe value, payments data}

\item{liquidity_provided}{dataframe value, optional parameter of a dataframe
containing the maximum net debit positions of all
participants. If not provided then this function
calculates these positions}
}
\value{
dataframe structure, daily gini coefficient of the system
}
\description{
Gini Coefficient
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to.
}
