% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liq_prov_rel_usage.R
\name{liq_prov_rel_usage}
\alias{liq_prov_rel_usage}
\title{Liquidity Provision Relative to Usage}
\usage{
liq_prov_rel_usage(participant, payments)
}
\arguments{
\item{participant}{character value, Identifying code of the payments system
participant}

\item{payments}{dataframe value, payments data}
}
\value{
dataframe structure, liquidity provision of the given participant
}
\description{
Liquidity Provision Relative to Usage
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
