% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liq_prov_rel_usage.R
\name{liq_prov_rel_usage}
\alias{liq_prov_rel_usage}
\title{Liquidity Provision Relative to Usage}
\usage{
liq_prov_rel_usage(payments)
}
\arguments{
\item{payments}{dataframe value, payments data}

\item{total_liquidity_provided}{dataframe value, daily system wide liquidity provided
defaults to NULL, the aggregate version of this
function calls the function with non NULL value}

\item{total_payments_sent}{dataframe value, daily payments sent
defaults to NULL, the aggregate version of this function
calls the function with non NULL value}
}
\value{
function, returns a closure that contains the aggregate level data
        in its environment. The closure takes one argument: participant name.
}
\description{
Liquidity Provision Relative to Usage
}
\details{
Assumes that the payments data file has the form:
ID, date, time, value, from, to
}
\examples{


participant_liq_prov <- liq_prov_rel_usage(example_data2)

# participant_liq_prov can now be called with a participants name
AAAAAA_liq_prov <- participant_liq_prov("AAAAAA")


}
